Get setTimeout lazily (#20)
* Get setTimeout and Date.now lazily

This prevents from using old versions of setTimeout/Date.now

* Modify withFakeTimers unit test to cover the correct scenario

BREAKING CHANGE: If you combined fake timers with wait-for-expect that will most probably not work anymore. Please make an issue with your failing usecase, as we couldn't think of a situation where using wait-for-expect and fake timers in the same test would be necessary.